(refactor) Inline address
