+++ async_dns/__init__.py
@@ -1,2 +1,2 @@
 from .base import *
-from . import types, hosts, address, utils
+from . import types, hosts, utils
